fix, slight project reorganization, added #if UNITY_EDITOR in a few places
