+++ Assets/Scripts/BuildingSystem/BuildingInfo.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Cosmobot.ItemSystem;
+using Cosmobot.Utils;
 using UnityEngine;
 
 namespace Cosmobot.BuildingSystem
