+++ Assets/Scripts/Editor/Cosmobot.Editor.asmdef
@@ -1,12 +1,12 @@
 {
-    "name": "Cosmobot",
-    "rootNamespace": "Cosmobot",
+    "name": "Cosmobot.Editor",
+    "rootNamespace": "Cosmobot.Editor",
     "references": [
-        "GUID:6055be8ebefd69e48b49212b09b47b2f",
-        "GUID:75469ad4d38634e559750d17036d5f7c",
         "GUID:d550786fe3d078e4b8037e8b66c4d019"
     ],
-    "includePlatforms": [],
+    "includePlatforms": [
+        "Editor"
+    ],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
