+++ Assets/Scripts/Enemies/TriggerAttackScript.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using Cosmobot.Entity;
+using Cosmobot.Utils;
 using UnityEngine;
 
 namespace Cosmobot
