+++ Assets/Scripts/Entity/HealthBar.cs
@@ -1,3 +1,4 @@
+using Cosmobot.Utils;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
