+++ Assets/Scripts/ItemSystem/Editor/ItemInspector.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
@@ -86,3 +87,4 @@ namespace Cosmobot.ItemSystem.Editor
         }
     }
 }
+#endif
