+++ Assets/Scripts/ItemSystem/ItemInfo.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Cosmobot.Utils;
 using UnityEngine;
 
 namespace Cosmobot.ItemSystem
