+++ Assets/Scripts/Utils/AnyUnityObjectUIAttribute.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Cosmobot
+namespace Cosmobot.Utils
 {
     // TODO: find a way to make SerializeReference work with UnityEngine.Object
     //   suggestion: Make field as UnityEngine.Object, use property drawer to check if the assigned object is of the
