+++ Assets/Scripts/Utils/Cosmobot.Utils.asmdef
@@ -1,11 +1,7 @@
 {
-    "name": "Cosmobot",
-    "rootNamespace": "Cosmobot",
-    "references": [
-        "GUID:6055be8ebefd69e48b49212b09b47b2f",
-        "GUID:75469ad4d38634e559750d17036d5f7c",
-        "GUID:d550786fe3d078e4b8037e8b66c4d019"
-    ],
+    "name": "Cosmobot.Utils",
+    "rootNamespace": "Cosmobot.Utils",
+    "references": [],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
