+++ Assets/Scripts/Utils/Cosmobot.Utils.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d550786fe3d078e4b8037e8b66c4d019
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
