+++ Assets/Tests/PlayModeTests/PlayModeTests.asmdef
@@ -4,7 +4,8 @@
     "references": [
         "UnityEngine.TestRunner",
         "UnityEditor.TestRunner",
-        "Cosmobot"
+        "Cosmobot",
+        "Cosmobot.Utils"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
